ui 수정 및 star rate 수정

+++ ParTner/app/src/main/java/com/example/partner/TrainerListRecyclerAdapter.java
@@ -42,7 +42,7 @@ public class TrainerListRecyclerAdapter extends RecyclerView.Adapter<TrainerList
     * */
     @Override
     public void onBindViewHolder(@NonNull MyViewHolder holder, final int position) {
-        float rate_score = listData.get(position).getStar_rate() / listData.get(position).getStar_rate_num();
+        float rate_score = (float)listData.get(position).getStar_rate() / (float)listData.get(position).getStar_rate_num();
         holder.ratingBar.setRating(rate_score);
         holder.name.setText(listData.get(position).getName());
         holder.profile.setText(listData.get(position).getSelf_introduction());
